Refactorer backend + sockets almost done
